Fix amount validation to ensure proper type conversion for net total comparison

+++ lib/app/pages/inventory/sales/create_sales/modals/sales_process_modal/sales_process_modal_controller.dart
@@ -277,7 +277,7 @@ class SalesProcessModalController extends PaymentGatewayController {
     final amountText = amountController.value.text;
     final isAmountEmpty = amountText.isEmptyOrNull;
     final amount = (num.tryParse(amountText) ?? 0).toInt();
-    final isInvalidAmount = amount >= 0 && amount < netTotal.value;
+    final isInvalidAmount = amount >= 0 && amount < netTotal.value.toInt();
 
     if (kDebugMode) {
       print('isZeroSalesAllowed: $isZeroSalesAllowed');
